+++ src/rules/functionality.ts
@@ -6,6 +6,8 @@
 import consoleMethods from './rule-helpers/console-methods';
 
 export default {
+  // Warns for an awaited value that is not a Promise.
+  'await-promise': true,
   // Bans the use of specific functions or global methods.
   'ban': false,
   // Enforces braces for `if`/`for`/`do`/`while` statements.
@@ -29,18 +31,24 @@ export default {
   'no-construct': true,
   // Disallows `debugger` statements.
   'no-debugger': true,
+   // Warns if ‘super()’ appears twice in a constructor.
+  'no-duplicate-super': true,
   // Disallows duplicate variable declarations in the same block scope.
   'no-duplicate-variable': true,
   // Disallows empty blocks.
   'no-empty': true,
   // Disallows `eval` function invocations.
   'no-eval': true,
+   // Promises returned by functions must be handled appropriately.
+  'no-floating-promises': false,
   // Disallows iterating over an array with a for-in loop.
   'no-for-in-array': true,
   // Disallow type inference of {} (empty object type) at function and constructor call sites
   'no-inferred-empty-object-type': true,
   // Disallows using the `this` keyword outside of classes.
   'no-invalid-this': true,
+  // Warns on apparent attempts to define constructors for interfaces or new for classes.
+  'no-misused-new': true,
   // Disallows use of the `null` keyword literal.
   'no-null-keyword': false,
   // Disallows shadowing variable declarations.
@@ -51,6 +59,10 @@ export default {
   'no-string-throw': true,
   // Disallows falling through case statements.
   'no-switch-case-fall-through': true,
+   // Warns when a method is used as outside of a method call.
+  'no-unbound-method': false,
+   // Warns when using an expression of type ‘any’ in an unsafe way. Type casts and tests are allowed. Expressions that work on all values (such as ‘”” + x’) are allowed.
+  'no-unsafe-any': false,
   // Disallows control flow statements, such as `return`, `continue`,
   // `break` and `throws` in finally blocks.
   'no-unsafe-finally': true,
@@ -71,6 +83,8 @@ export default {
   // Usage of &amp;&amp; or || operators should be with boolean operands and
   // expressions in If, Do, While and For statements should be of type boolean
   'strict-boolean-expressions': false,
+  // Warns for type predicates that are always true or always false. Works for ‘typeof’ comparisons to constants (e.g. ‘typeof foo === “string”’), and equality comparison to ‘null’/’undefined’. (TypeScript won’t let you compare ‘1 === 2’, but it has an exception for ‘1 === undefined’.) Does not yet work for ‘instanceof’. Does not warn for ‘if (x.y)’ where ‘x.y’ is always truthy. For that, see strict-boolean-expressions.
+  'strict-type-predicates': false,
   // Require a `default` case in all `switch` statements.
   'switch-default': false,
   // Requires `===` and `!==` in place of `==` and `!=`.